Fixed setup, added entry point, added README.rst.
Setup was missing a requirement, but it didn't really need it.  Rather
than installing fonts into the user's home directory, we can just
reference them in the package.

I've converted the README.md to a README.rst using pandoc:

    pandoc README.md -o README.rst

since pypi uses the rst format.

I've also added an entry point to the setup file, so that the command
line utility can be used without knowing where it is located:

    picol --help